+++ permissions/src/epilot/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.2.3'
-    gen_version: str = '2.237.2'
-    user_agent: str = 'speakeasy-sdk/python 2.2.3 2.237.2 1.0.0 epilot-permissions'
+    sdk_version: str = '2.2.4'
+    gen_version: str = '2.245.1'
+    user_agent: str = 'speakeasy-sdk/python 2.2.4 2.245.1 1.0.0 epilot-permissions'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: